+++ CHANGES.rst
@@ -21,6 +21,8 @@ Improvements
 ^^^^^^^^^^^^
 
 - Invalidate password reset links once the password has been changed (:pr:`5878`)
+- Add full ACLs for custom conference menu items, instead of just being able to
+  restrict them to speakers or registrants (:pr:`5670`, thanks :user:`kewisch`)
 
 Bugfixes
 ^^^^^^^^